Fix clang and gcc warnings (#552)
* fix lookAndFeel shadowing warnings + others, update algogris

* more warning fixes

* more warning fixes

* update algogris

* update algogris

* msvc doesn't support constexpr fpclassify yet

* Update AlgoGRIS

---------

+++ Source/sg_SpeakerViewComponent.cpp
@@ -482,7 +482,7 @@ void SpeakerViewComponent::listenUDP(juce::DatagramSocket& socket)
     auto packetSize = socket.read(receiveBuffer, mMaxBufferSize, false, senderAddress, senderPort);
 
     if (packetSize > 0) {
-        juce::String receivedData(receiveBuffer, packetSize);
+        juce::String receivedData(receiveBuffer, static_cast<size_t>(packetSize));
         juce::var jsonResult;
         auto res = juce::JSON::parse(receivedData, jsonResult);
